+++ meteoio/meteostats/libinterpol2D.h
@@ -63,7 +63,7 @@ class Interpol2D {
                                 const DEMObject& dem, Grid2DObject& grid);
 		static void LocalLapseIDW(const std::vector<double>& vecData_in,
 		                          const std::vector<StationData>& vecStations_in,
-		                          const DEMObject& dem, const unsigned int& nrOfNeighbors,
+		                          const DEMObject& dem, const size_t& nrOfNeighbors,
 		                          Grid2DObject& grid, double& r2);
 		static void SimpleDEMWindInterpolate(const DEMObject& dem, Grid2DObject& VW, Grid2DObject& DW);
 		static void PrecipSnow(const DEMObject& dem, const Grid2DObject& ta, Grid2DObject& grid);
@@ -95,7 +95,7 @@ class Interpol2D {
 		static double getReferenceAltitude(const DEMObject& dem);
 		static void getNeighbors(const double& x, const double& y,
 		                         const std::vector<StationData>& vecStations,
-		                         std::vector< std::pair<double, unsigned int> >& list);
+		                         std::vector< std::pair<double, size_t> >& list);
 
 		//core methods
 		static double IDWCore(const double& x, const double& y,
@@ -104,7 +104,7 @@ class Interpol2D {
 		static double LLIDW_pixel(const unsigned int& i, const unsigned int& j,
 		                          const std::vector<double>& vecData_in,
 		                          const std::vector<StationData>& vecStations_in,
-		                          const DEMObject& dem, const unsigned int& nrOfNeighbors, double& r2);
+		                          const DEMObject& dem, const size_t& nrOfNeighbors, double& r2);
 
 		//weighting methods
 		static double weightInvDist(const double& d2);
